src/tools/crawler_test_function/requirements.txt
google-cloud-pubsub == 2.2.0
google-cloud-bigquery == 2.7.0
google-cloud-storage == 1.35.0
google-api-python-client == 2.0.2
google-cloud-secret-manager == 2.4.0
google-cloud-logging
oauth2client == 4.1.3
lxml == 4.6.5
selenium == 3.13.0
SQLAlchemy == 1.3.12
fake-useragent==0.1.11
pandas
webdriver-manager == 3.8.2
pyOpenSSL == 22.0.0
python-dateutil == 2.8.2
boto3
black
beautifulsoup4
